devmem: Allow for easy printing of blocks of hardware registers
Sometimes when reverse engineering a piece of undocumented hardware
it is useful to boot a vendor version of u-boot, the kernel etc
dump the register values and then boot a vanilla version of the
same thing, dump the register values again and then compare them
to find the needed magic values, undocumented bits etc.

Tools like hexdump don't seem to be able to access anything
but normal memory via /dev/mem so devmem seems to be the only
option.

Devmem is great for this except that it can only dump out one
register at a time. So you need a script to dump out a whole region
and you probably have to handle converting between bases etc in your
little script.

So to save my sanity allow devmem to take a count of the number
of units the user wants to read or write.

The syntax is a little weird but shouldn't break anyones existing
hacks using devmem.